requirements.txt
absl-py==2.3.1
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.12.1
altair==4.2.2
altgraph @ file:///AppleInternal/Library/BuildRoots/4~B7e5ugCABw73MAeFSK0ogoFHGg4yzTtKi73iQ2w/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==3.7.1
anywidget==0.9.18
asgiref==3.9.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
base58==2.1.1
bcrypt==4.2.0
bitarray==3.7.1
blinker==1.9.0
bqplot==0.12.45
branca==0.8.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
choreographer==1.1.1
ckzg==1.0.2
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
comm==0.2.3
config==0.5.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
cytoolz==1.0.1
decorator==5.2.1
Django==4.2.24
dnspython==2.7.0
earthengine-api==0.1.370
ecdsa==0.19.1
eerepr==0.1.2
email-validator==2.3.0
entrypoints==0.4
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.104.1
filelock==3.19.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
folium==0.20.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
future @ file:///AppleInternal/Library/BuildRoots/4~B7e5ugCABw73MAeFSK0ogoFHGg4yzTtKi73iQ2w/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gast==0.6.0
geemap==0.36.2
geocoder==1.38.1
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hexbytes==0.3.1
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.25.2
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipfshttpclient==0.7.0
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.20.0
ipython==8.18.1
ipytree==0.2.2
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-leaflet==0.20.0
jupyterlab_widgets==3.0.15
kagglehub==0.3.13
kaleido==1.1.0
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
logistro==1.1.0
lru-dict==1.2.0
lxml==6.0.2
macholib @ file:///AppleInternal/Library/BuildRoots/4~B7e5ugCABw73MAeFSK0ogoFHGg4yzTtKi73iQ2w/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Mako==1.3.10
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.6.4
mysql-connector-python==9.4.0
mysqlclient==2.2.7
namex==0.1.0
narwhals==2.3.0
netaddr==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
packaging==23.2
pandas==2.0.3
parsimonious==0.10.0
parso==0.8.5
passlib==1.7.4
pexpect==4.9.0
Pillow==9.5.0
pinatapy-vourhey==0.2.0
plotly==6.3.0
pluggy==1.6.0
polars==1.33.1
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.1.0
psycopg2-binary==2.9.9
psygnal==0.14.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-ecc==8.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.0
Pympler==1.1
PyMySQL==1.1.2
pyogrio==0.11.1
pyparsing==3.2.3
pyperclip==1.9.0
pyproj==3.6.1
pyshp==2.3.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.4.0
python-box==7.3.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.4
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pyunormalize==16.0.0
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.0
rasterio==1.3.10
ratelim==0.1.6
referencing==0.36.2
regex==2025.9.1
reportlab==4.2.2
requests==2.31.0
rich==13.9.4
rlp==4.1.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.13.1
scooby==0.10.1
shapely==2.0.3
simplejson==3.20.1
six @ file:///AppleInternal/Library/BuildRoots/4~B7e5ugCABw73MAeFSK0ogoFHGg4yzTtKi73iQ2w/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.2
sniffio==1.3.1
snuggs==1.4.7
SQLAlchemy==2.0.23
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.27.0
streamlit==1.50.0
streamlit-cropper==0.2.2
streamlit-drawable-canvas==0.9.3
streamlit-image-comparison==0.0.4
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==4.3.1
ultralytics==8.3.202
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
validators==0.35.0
varint==1.0.2
watchfiles==1.1.0
wcwidth==0.2.13
web3==6.19.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.3
xyzservices==2025.4.0
yarl==1.20.1
zipp==3.23.0
